split out the archive-building specific logic from the CourseAnalyticsExport and into a dedicated Analytics::ExportBuilder class
